Shulker size fix, velocityChanged makes the entity velocity reset, so it was removed
I have no other idea how to fix the jitter glitch. I'm pissed to have made it this far and be stumped on an effect that takes just a few lines of code. I spent two full days with no solution and watching the test villager do the same exact jitter.

Turns out the client and server potion effects are not properly synced, so that causes it. For now, it'll just jitter I guess unless anyone has a working solution that doesn't reset the entity's velocity, I give up for now...