Fix build failure because IntelliJ apparently has trouble showing build errors...
And just like that, I went from thinking this IDE was 10/10, to 0/10 because apparently IntelliJ thinks it's a good idea to WAIT UNTIL YOU OPEN A FILE TO SHOW YOU A FUCKING COMPILE ERROR. And the sane behavior is FUCKING DEFAULT TO OFF!?

but if I go back to Eclipse I gotta deal with slow BS, no mod dev plugin and files randomly not saving. Why does no fucking functioning IDE exist?