Add CraftTweaking for brewing fuel, enchanting fuel, composting
Also removed IBrewingFuel because it seems to be an extreme edgecase when it comes to uses. It's only useful for if someone wants nbt-sensitive values for different brew counts based on NBT, and allows for multiple items in one fuel class which makes it cumbersome to check for and integrate into CT.
So I removed it.

Enchanting tables now show random item names from the compatible fuels list, and will dispense the container item if the fuel has a container item.

Brewing fuel's container item implementation was left unchanged, we ignore doesContainerItemLeaveCraftingGrid because the consumption of the brewing item is not based directly on player input.