Fix Elytra DataWatcher, partial ASJCore incompatibility
Use proper fetching and setting of Data Watcher flags as well as making it configurable.

Also remove unused shadowing of "flyToggleTimer"

Fixes part of #216, ASJCore still needs to fix the furnace issue where it adds an invalid recipe.

Yes I could add a null check but since mods should not add null recipes to the furnace it's a good way to see what mods need to be fixed. I will finally add a null check if an ARR or otherwise unmaintained mod ends up causing it.