Dynamically generate sounds.json
Just contains cave sounds for now, might add more in the future.
This doesn't work for some reason. I know the game is reading this dynamically generated sounds.json but it seems to make the ambient.cave.cave event not play anything sometimes and doesn't add any of the new sounds to it.
Logger left in on purpose to demonstrate the json works correctly